+++ src/main/java/com/account/yomankum/security/oauth/filter/CustomOAuth2AuthorizationCodeGrantFilter.java
@@ -49,40 +49,38 @@ public class CustomOAuth2AuthorizationCodeGrantFilter extends OAuth2Authorizatio
         }
 
         String requestURI = request.getRequestURL().toString();
-        HttpSession session = request.getSession(false);
+        HttpSession session = request.getSession();
 
-        if (session != null) {
-            String myState = String.valueOf(
-                    session.getAttribute(TokenProp.STATE.getName())
-            );
-            String snsSendState = request.getParameter(TokenProp.STATE.getName());
+        String myState = String.valueOf(
+                session.getAttribute(TokenProp.STATE.getName())
+        );
+        String snsSendState = request.getParameter(TokenProp.STATE.getName());
 
-            if (myState.equals("null") | snsSendState == null) {
-                log.error("{} 없음.", TokenProp.STATE.getName());
-                filterChain.doFilter(request, response);
-                return;
-            }
-
-            if (snsSendState.equals(myState) && StringUtils.hasText(code)) {
-                // state 값이 key 고 value 가 sns 명
-                String sns = String.valueOf(session.getAttribute(snsSendState));
-                String clientId = snsInfo.getClientId();
-                String clientSecret = snsInfo.getClientSecret();
-                String tokenUri = snsInfo.getTokenUri();
-
-                removeSessionAttributeState(session, myState);
-
-                HttpEntity<MultiValueMap<String, String>> httpEntity = setHttpEntity
-                        (code, requestURI, sns, clientId, clientSecret);
-                // https://kauth.kakao.com/oauth/token 으로 토큰 요청 보내기
-                ResponseEntity<TokenResponse> responseEntity = sendTokenRequest(tokenUri, httpEntity);
-
-
-                TokenResponse tokenResponse = responseEntity.getBody();
-                request.setAttribute(TokenProp.TOKEN_RESPONSE.name(), tokenResponse);
-                request.setAttribute(TokenProp.SNS.getName(), sns);
-            }
+        if (myState.equals("null") | snsSendState == null) {
+            log.error("{} 없음.", TokenProp.STATE.getName());
+            filterChain.doFilter(request, response);
+            return;
         }
+
+        if (snsSendState.equals(myState) && StringUtils.hasText(code)) {
+            // state 값이 key 고 value 가 sns 명
+            String sns = String.valueOf(session.getAttribute(snsSendState));
+            String clientId = snsInfo.getClientId();
+            String clientSecret = snsInfo.getClientSecret();
+            String tokenUri = snsInfo.getTokenUri();
+
+            removeSessionAttributeState(session, myState);
+
+            HttpEntity<MultiValueMap<String, String>> httpEntity = setHttpEntity
+                    (code, requestURI, sns, clientId, clientSecret);
+            // https://kauth.kakao.com/oauth/token 으로 토큰 요청 보내기
+            ResponseEntity<TokenResponse> responseEntity = sendTokenRequest(tokenUri, httpEntity);
+
+            TokenResponse tokenResponse = responseEntity.getBody();
+            request.setAttribute(TokenProp.TOKEN_RESPONSE.name(), tokenResponse);
+            request.setAttribute(TokenProp.SNS.getName(), sns);
+        }
+
 
         filterChain.doFilter(request, response);
     }
