+++ src/main/java/com/account/yomankum/security/oauth/token/JwtValue.java
@@ -2,13 +2,10 @@ package com.account.yomankum.security.oauth.token;
 
 public interface JwtValue {
 
-    String getFirstKid();
-    String getSecondKid();
     String getKid();
     String getKty();
     String getAlg();
     String getUse();
     String getN();
     String getE();
-    void jwkSetting(String number);
 }
